+++ lib/dotsync.rb
@@ -7,25 +7,23 @@ require "toml-rb"
 require "logger"
 require "forwardable" # Ruby standard library
 require "ostruct"
-
-# Errors
-require_relative "dotsync/errors"
+require "find"
 
 # Utils
-require_relative "dotsync/icons"
-require_relative "dotsync/logger"
-require_relative "dotsync/file_transfer"
-require_relative "dotsync/path_utils"
+require_relative "dotsync/utils/logger"
+require_relative "dotsync/utils/file_transfer"
+require_relative "dotsync/utils/directory_differ"
+require_relative "dotsync/utils/path_utils"
 
 # Config
-require_relative "dotsync/actions/config/xdg_base_directory_spec"
-require_relative "dotsync/actions/config/mapping_entry"
-require_relative "dotsync/actions/config/base_config"
-require_relative "dotsync/actions/config/pull_action_config"
-require_relative "dotsync/actions/config/push_action_config"
-require_relative "dotsync/actions/config/watch_action_config"
+require_relative "dotsync/config/xdg_base_directory_spec"
+require_relative "dotsync/config/mapping_entry"
+require_relative "dotsync/config/base_config"
+require_relative "dotsync/config/pull_action_config"
+require_relative "dotsync/config/push_action_config"
+require_relative "dotsync/config/watch_action_config"
 
-# Concerns
+# Actions Concerns
 require_relative "dotsync/actions/concerns/mappings_transfer"
 
 # Actions
@@ -34,13 +32,13 @@ require_relative "dotsync/actions/pull_action"
 require_relative "dotsync/actions/push_action"
 require_relative "dotsync/actions/watch_action"
 
+# Base classes
+require_relative "dotsync/errors"
+require_relative "dotsync/icons"
 require_relative "dotsync/runner"
-
 require_relative "dotsync/version"
 
 module Dotsync
-  class Error < StandardError; end
-
   class << self
     attr_writer :config_path