+++ lib/dotsync/errors.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
 module Dotsync
+  class Error < StandardError; end
   class ConfigError < StandardError; end
 end