+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 0.1.5
 
 - Fixes backup when destination folder does not exist
-- Add rubocop
+- Add rubocop to Github Actions
 
 
 # 0.1.4